reformat as a program instead of a script
