--- dns_structures.py
@@ -1,239 +1,0 @@
-from impacket.structure import Structure
-import socket
-from struct import unpack, pack
-import datetime
-
-# mostly stolen from adidnsdump, specifically: https://github.com/dirkjanm/adidnsdump/blob/65169b2b5c9dc2b51afe03851d42156085c4cd68/adidnsdump/dnsdump.py#L327
-
-# https://en.wikipedia.org/wiki/List_of_DNS_record_types
-RECORD_TYPE_MAPPING = {
-    0: 'ZERO',
-    1: 'A',
-    2: 'NS',
-    3: 'MD',
-    4: 'MF',
-    5: 'CNAME',
-    6: 'SOA',
-    7: 'MB',
-    8: 'MG',
-    9: 'MR',
-    10: 'NULL',
-    11: 'WKS',
-    12: 'PTR',
-    13: 'HINFO',
-    14: 'MINFO',
-    15: 'MX',
-    16: 'TXT',
-    17: 'RP',
-    18: 'AFSDB',
-    19: 'X25',
-    20: 'ISDN',
-    21: 'RT',
-    24: 'SIG',
-    25: 'KEY',
-    28: 'AAAA',
-    30: 'NXT',
-    33: 'SRV',
-    34: 'ATMA',
-    35: 'NAPTR',
-    39: 'DNAME',
-    43: 'DS',
-    46: 'RRSIG',
-    47: 'NSEC',
-    48: 'DNSKEY',
-    49: 'DHCID',
-    50: 'NSEC3',
-    51: 'NSEC3PARAM',
-    52: 'TLSA',
-    65281: 'WINS', # this may be a negative number its xFF01 (?)
-    65281: 'WINSR', # ^^^
-}
-
-
-class DNS_RECORD(Structure):
-    """
-    dnsRecord - used in LDAP
-    [MS-DNSP] section 2.3.2.2
-    """
-    structure = (
-        ('DataLength', '<H-Data'),
-        ('Type', '<H'),
-        ('Version', 'B=5'),
-        ('Rank', 'B'),
-        ('Flags', '<H=0'),
-        ('Serial', '<L'),
-        ('TtlSeconds', '>L'),
-        ('Reserved', '<L=0'),
-        ('TimeStamp', '<L=0'),
-        ('Data', ':')
-    )
-
-# Note that depending on whether we use RPC or LDAP all the DNS_RPC_XXXX
-# structures use DNS_RPC_NAME when communication is over RPC,
-# but DNS_COUNT_NAME is the way they are stored in LDAP.
-#
-# Since LDAP is the primary goal of this script we use that, but for use
-# over RPC the DNS_COUNT_NAME in the structures must be replaced with DNS_RPC_NAME,
-# which is also consistent with how MS-DNSP describes it.
-
-class DNS_RPC_NAME(Structure):
-    """
-    DNS_RPC_NAME
-    Used for FQDNs in RPC communication.
-    MUST be converted to DNS_COUNT_NAME for LDAP
-    [MS-DNSP] section 2.2.2.2.1
-    """
-    structure = (
-        ('cchNameLength', 'B-dnsName'),
-        ('dnsName', ':')
-    )
-
-class DNS_COUNT_NAME(Structure):
-    """
-    DNS_COUNT_NAME
-    Used for FQDNs in LDAP communication
-    MUST be converted to DNS_RPC_NAME for RPC communication
-    [MS-DNSP] section 2.2.2.2.2
-    """
-    structure = (
-        ('Length', 'B-RawName'),
-        ('LabelCount', 'B'),
-        ('RawName', ':')
-    )
-
-    def toFqdn(self):
-        ind = 0
-        labels = []
-        for i in range(self['LabelCount']):
-            nextlen = unpack('B', self['RawName'][ind:ind+1])[0]
-            labels.append(self['RawName'][ind+1:ind+1+nextlen].decode('utf-8'))
-            ind += nextlen + 1
-        # For the final dot
-        labels.append('')
-        return '.'.join(labels)
-
-class DNS_RPC_NODE(Structure):
-    """
-    DNS_RPC_NODE
-    [MS-DNSP] section 2.2.2.2.3
-    """
-    structure = (
-        ('wLength', '>H'),
-        ('wRecordCount', '>H'),
-        ('dwFlags', '>L'),
-        ('dwChildCount', '>L'),
-        ('dnsNodeName', ':')
-    )
-
-class DNS_RPC_RECORD_A(Structure):
-    """
-    DNS_RPC_RECORD_A
-    [MS-DNSP] section 2.2.2.2.4.1
-    """
-    structure = (
-        ('address', ':'),
-    )
-
-    def formatCanonical(self):
-        return socket.inet_ntoa(self['address'])
-
-    def fromCanonical(self, canonical):
-        self['address'] = socket.inet_aton(canonical)
-
-
-class DNS_RPC_RECORD_NODE_NAME(Structure):
-    """
-    DNS_RPC_RECORD_NODE_NAME
-    [MS-DNSP] section 2.2.2.2.4.2
-    """
-    structure = (
-        ('nameNode', ':', DNS_COUNT_NAME),
-    )
-
-class DNS_RPC_RECORD_SOA(Structure):
-    """
-    DNS_RPC_RECORD_SOA
-    [MS-DNSP] section 2.2.2.2.4.3
-    """
-    structure = (
-        ('dwSerialNo', '>L'),
-        ('dwRefresh', '>L'),
-        ('dwRetry', '>L'),
-        ('dwExpire', '>L'),
-        ('dwMinimumTtl', '>L'),
-        ('namePrimaryServer', ':', DNS_COUNT_NAME),
-        ('zoneAdminEmail', ':', DNS_COUNT_NAME)
-    )
-
-class DNS_RPC_RECORD_NULL(Structure):
-    """
-    DNS_RPC_RECORD_NULL
-    [MS-DNSP] section 2.2.2.2.4.4
-    """
-    structure = (
-        ('bData', ':'),
-    )
-
-
-class DNS_RPC_RECORD_STRING(Structure):
-    """
-    DNS_RPC_RECORD_NODE_NAME
-    [MS-DNSP] section 2.2.2.2.4.6
-    """
-    structure = (
-        ('stringData', ':', DNS_COUNT_NAME),
-    )
-
-
-# Some missing structures here that I skipped
-
-class DNS_RPC_RECORD_NAME_PREFERENCE(Structure):
-    """
-    DNS_RPC_RECORD_NAME_PREFERENCE
-    [MS-DNSP] section 2.2.2.2.4.8
-    """
-    structure = (
-        ('wPreference', '>H'),
-        ('nameExchange', ':', DNS_COUNT_NAME)
-    )
-
-# Some missing structures here that I skipped
-
-class DNS_RPC_RECORD_AAAA(Structure):
-    """
-    DNS_RPC_RECORD_AAAA
-    [MS-DNSP] section 2.2.2.2.4.17
-    """
-    structure = (
-        ('ipv6Address', '16s'),
-    )
-
-    def formatCanonical(self):
-        return socket.inet_ntop(socket.AF_INET6, self['ipv6Address'])
-
-class DNS_RPC_RECORD_SRV(Structure):
-    """
-    DNS_RPC_RECORD_SRV
-    [MS-DNSP] section 2.2.2.2.4.18
-    """
-    structure = (
-        ('wPriority', '>H'),
-        ('wWeight', '>H'),
-        ('wPort', '>H'),
-        ('nameTarget', ':', DNS_COUNT_NAME)
-    )
-
-class DNS_RPC_RECORD_TS(Structure):
-    """
-    DNS_RPC_RECORD_TS
-    [MS-DNSP] section 2.2.2.2.4.23
-    """
-    structure = (
-        ('entombedTime', '<Q'),
-    )
-    def toDatetime(self):
-        microseconds = int(self['entombedTime'] / 10)
-        try:
-            return datetime.datetime(1601,1,1) + datetime.timedelta(microseconds=microseconds)
-        except OverflowError:
-            return None
